refactor: at Search component get categories from api + refactor: centralize constants in a single file + feat: add placeholderBlurParams constant
