+++ src/components/CartItem.tsx
@@ -4,7 +4,7 @@ import { formatPrice } from "@/utils/formatPrice";
 import Image from "next/image";
 import React from "react";
 import CustomNumberInput from "./CustomNumberInput";
-import { navigation } from "@/constants/navigation";
+import { navigation } from "@/constants";
 import Link from "next/link";
 
 export default function CartItem({
