+++ src/components/FabricDescription.tsx
@@ -5,6 +5,7 @@ import { motion, useInView } from "framer-motion";
 import Image from "next/image";
 import { Divider } from "antd";
 import { EyeOutlined } from "@ant-design/icons";
+import { placeholderBlurParams } from "@/constants";
 
 interface IProps {
   fabric: Fabric;
@@ -94,10 +95,7 @@ export default function FabricDescription({
                   height={100}
                   className="object-contain object-center"
                   placeholder="blur"
-                  blurDataURL={
-                    c.foto +
-                    "?w=20&h=20&q=50&fm=webp&fit=crop&blur=500&px=16&auto=format"
-                  }
+                  blurDataURL={c.foto + placeholderBlurParams}
                 />
                 <div className="opacity-0 group-hover:opacity-100 transition-opacity duration-75 absolute h-full w-full bg-black/50 flex-row-center top-0">
                   <EyeOutlined style={{ color: "white", fontSize: 18 }} />
