--- src/constants/categories.ts
@@ -1,8 +1,0 @@
-
-export const categories: string[] = [
-    'silla',
-    'sillón',
-    'sofá',
-    'puff',
-    'mesa'
-]
